requirements.txt
Django==1.10.1
djangorestframework==3.4.6
django-extensions==1.7.4
psycopg2==2.6.2
